One Shot TImer and Timer Duration Correction
To get the one shot timer to work, a software triggered UEV was performed by setting the UG in TIMx_EGR,
after the prescaler and ARR value had been set. This would trigger the system to update in the pre-
scale counter and counter register. Afterwards clear the UIF flag in TIMx_SR and enable CEN in TIMx_CR1.

With regards to Timer 2-5 is connected to the APB1 which beats at 42MHz. If TIMPRE is reset in RCC_DCKCFGR,
the TIMx_CLK = 2 x PCLK if the prescale for APB1 is > 1 else TIMx_CLK = PCLK.